refactoring 1D code
