+++ analysis_tools/TH/oned_flow.py
@@ -10,15 +10,9 @@ import copy
 import numpy as np
 
 # define global variables
-D = 0.01 # fuel diameter [m]
-PD = 1.1 # pitch to diameter ratio [-] 
+D = 0.01 # flow channel diameter [m]
+PD = 1.05 # pitch to diameter ratio [-] 
 L = 0.5 # fuel length [m]
-k_c = 21e-3 # clad conductivity [kW/m-k]
-k_f = 21e-3 # fuel conductivity [kW/m-k]
-k_cool = 0.07527e-3 # coolant conductivity [kW/m-k], choose min of inlet/outlet k
-nu = 0.00004305 # coolant viscosity [kg/m-s], chose min of inlet/outlet
-Cp_cool = 1.274 # coolant specific heat [kJ/kg-k], choose min of in/out
-rho_cool = 85.99 # coolant density [kg/m^3], choose min of in/out
 c = 0.00031 # clad thickness [m] 
 Q_therm = 131.2 # core thermal power [kW]
 T_in = 962.9 # core inlet temp [K] (from power cycle model)
@@ -46,23 +40,6 @@ def get_q_bar(R, h, dt, h_bar_switch=1):
      
     return q_bar, t_clad, q_trip
 
-def get_mass_flux_square(PD, N_elements):
-    """Calculate coolant mass flux for a square lattice.
-    Arguments: PD (float) pitch-to-diameter ratio
-               N_elements (int) number of fuel elements
-               D (float): diameter of fuel pin
-               m_dot (float): core mass flow rate.
-    Returns: G_dot (float): coolant mass flux
-    """
-    pitch_area = (PD * D) ** 2.0 
-    pin_area = math.pi * (D / 2.0) ** 2.0
-    flow_area = ((pitch_area - pin_area) / 1e4) * N_elements
-    flow_perim = math.pi * D * N_elements
-    G_dot = m_dot / flow_area
-    D_e = 4 * flow_area / flow_perim
-    v = G_dot / rho_cool
-
-    return G_dot, D_e, v
 
 def get_mass_flux_triangle(PD, N_elements):
     """Calculate coolant mass flux for a triangle lattice.
@@ -84,21 +61,6 @@ def get_mass_flux_triangle(PD, N_elements):
     
     return G_dot, D_e, v
 
-def get_h_bar(G_dot, D_e):
-    """Calculate average heat transfer coefficient.
-    Arguments: 
-    Returns: h_bar (float) average heat transfer coefficient.
-    """
-    
-    Re = D_e * G_dot/ nu
-    Pr = Cp_cool * nu / k_cool
-    
-    # equation is combination of 9-30, 9-31b in El-Wakil's Nuclear Heat
-    # Transport. Valid for 1.1 <= PD <= 1.3
-    Nu = (0.026 * PD - 0.006) * math.pow(Re, 0.8) * math.pow(Pr, 0.333)
-    h_bar = Nu * k_cool / D_e
-    
-    return h_bar, Re
 
 def calc_dp(D_e, v, Re):
     """Calculate pressure drop in triangle-pitch subchannel
@@ -121,7 +83,7 @@ def run_iter(PD, diam=D):
     # get first estimate for q_bar
     q_bar, t_clad, q_trip = get_q_bar(r_f, 1.0, dt, 0)
     N_pins = math.ceil(Q_therm / q_bar)
-    G_dot, D_e, v = get_mass_flux_triangle(PD, N_pins)
+    G_dot, D_e, v = get_mass_flux_triangle(D/2, N_pins)
     h_bar, Re = get_h_bar(G_dot, D_e)
 
     q_bar_new = get_q_bar(r_f, h_bar, dt, 1)[0]
